Feature: Add initial utils and their usage
