+++ core/domain/src/main/java/one/reevdev/stood/core/domain/task/model/Category.kt
@@ -1,7 +1,9 @@
 package one.reevdev.stood.core.domain.task.model
 
+import one.reevdev.cosmoe.utils.emptyString
+
 data class Category(
-    val id: String,
-    val name: String,
-    val color: String,
+    val id: String = emptyString(),
+    val name: String = emptyString(),
+    val color: String = "#000000",
 )
