+++ cosmoe/src/main/java/one/reevdev/cosmoe/ui/compose/UiState.kt
@@ -1,4 +1,4 @@
-package one.reevdev.stood.features.task.utils
+package one.reevdev.cosmoe.ui.compose
 
 interface UiState {
     val isLoading: Boolean
