+++ cosmoe/src/main/java/one/reevdev/cosmoe/utils/CommonUtils.kt
@@ -1,0 +1,7 @@
+package one.reevdev.cosmoe.utils
+
+fun <T> T?.orDefault(defaultValue: T) = this ?: defaultValue
+
+fun <T> T?.isNull() = this == null
+
+fun <T> T?.isNotNull() = this != null
