+++ cosmoe/src/main/java/one/reevdev/cosmoe/utils/StringUtils.kt
@@ -1,0 +1,5 @@
+package one.reevdev.cosmoe.utils
+
+fun emptyString() = ""
+
+fun String.isEmpty() = !this.isNotEmpty()
